requirements.txt
arabic-reshaper==2.1.4
asgiref==3.6.0
asn1crypto==1.5.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.4
django-gsheets==0.0.10
django-import-export==3.1.0
django-jet-reboot==1.3.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gunicorn==20.0.4
html5lib==1.1
httplib2==0.22.0
idna==3.4
lxml==4.9.2
MarkupPy==1.14
numpy==1.24.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
pandas==1.5.3
Pillow==9.3.0
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
supervisor==4.2.5
svglib==1.4.1
tablib==3.4.0
tinycss2==1.2.1
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.13
uuid==1.30
webencodings==0.5.1
whitenoise==6.3.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
